pyproject.toml
[tool.poetry]
name = "pd-ai"
version = "0.1.0"
description = ""
authors = ["Raphael Oliveira <raphael.foliveira1@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic-ai = "^0.0.29"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.2.5"}
psycopg-pool = "^3.2.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
